fixed tab navigation to remove extra tabs, added 'add item' page to proper place
