+++ ProdUSE/app.json
@@ -5,7 +5,7 @@
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/images/icon.png",
-    "scheme": "myapp",
+    "scheme": "ProdUSE",
     "userInterfaceStyle": "automatic",
     "splash": {
       "image": "./assets/images/splash.png",
