+++ ProdUSE/app/(tabs)/pantry.tsx
