+++ ProdUSE/app/_layout.tsx
@@ -1,4 +1,4 @@
-import { Tabs } from 'expo-router';
+import { Stack } from 'expo-router';
 import React from 'react';
 import { TabBarIcon } from '@/components/navigation/TabBarIcon';
 import { Colors } from '@/constants/Colors';
@@ -9,72 +9,33 @@ import DbMigrationRunner from '../db/DbMigrationRunner';
 import SQLiteProvider from '../db/SQLiteProvider';
 
 
-export default function TabLayout() {
-  const colorScheme = useColorScheme();
-  const [ready, setReady] = React.useState(false);
-  const migrateDbIfNeeded = async (db: SQLiteDatabase) => {
-    try {
-      await new DbMigrationRunner(db).apply(migrations);
-      console.log('All migrations applied.');
-      setReady(true);
-    } catch (err) {
-      console.error(err);
-    }
-  };
+const RootLayout = () => {
+    const colorScheme = useColorScheme();
+    const [ready, setReady] = React.useState(false);
+    const migrateDbIfNeeded = async (db: SQLiteDatabase) => {
+        try {
+        await new DbMigrationRunner(db).apply(migrations);
+        console.log('All migrations applied.');
+        setReady(true);
+        } catch (err) {
+        console.error(err);
+        }
+    };
 
   return (
     <SQLiteProvider
       databaseName = "produsedb.db"
       onInit = {migrateDbIfNeeded}>
-        {ready && 
-          <Tabs
-          screenOptions={{
-            tabBarActiveTintColor: 'green', 
-            tabBarInactiveTintColor: 'gray', 
-            headerShown: false,
-            tabBarStyle: {
-              marginTop: 10, 
-            },
-            }}
-          >
-            <Tabs.Screen
-              name="index" 
-              options={{
-                title: 'My Fridge',
-                tabBarIcon: ({ color, focused }) => (
-                  <TabBarIcon name={focused ? 'home' : 'home-outline'} color={color} />
-                ),
-              }}
+        {ready && <Stack>
+            <Stack.Screen
+                name="(tabs)"
+                options={{
+                    headerShown: false
+                }}
             />
-            <Tabs.Screen
-              name="pantry" 
-              options={{
-                title: 'pantry',
-                tabBarIcon: ({ color, focused }) => (
-                  <TabBarIcon name={focused ? 'basket' : 'basket-outline'} color={color} />
-                ),
-              }}
-            />
-            <Tabs.Screen
-              name="metrics" 
-              options={{
-                title: 'metrics',
-                tabBarIcon: ({ color, focused }) => (
-                  <TabBarIcon name={focused ? 'analytics' : 'analytics-outline'} color={color} />
-                ),
-              }}
-            />
-            <Tabs.Screen
-              name="profile" 
-              options={{
-                title: 'profile',
-                tabBarIcon: ({ color, focused }) => (
-                  <TabBarIcon name={focused ? 'person' : 'person-outline'} color={color} />
-                ),
-              }}
-            />
-          </Tabs>
-         }
+        </Stack>}
     </SQLiteProvider>
-  );
+    )
 }
+
+export default RootLayout
