+++ ProdUSE/package.json
@@ -17,27 +17,32 @@
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
     "@react-native-community/datetimepicker": "8.2.0",
+    "@react-navigation/bottom-tabs": "^7.0.0",
     "@react-navigation/native": "^7.0.0",
-    "expo": "^52.0.7",
+    "@react-navigation/native-stack": "^7.0.0",
+    "@types/react-dom": "^18.3.1",
+    "@types/scheduler": "^0.23.0",
+    "expo": "~52.0.11",
     "expo-constants": "~17.0.3",
     "expo-font": "~13.0.1",
-    "expo-linking": "~7.0.2",
-    "expo-router": "~4.0.5",
-    "expo-splash-screen": "~0.29.9",
+    "expo-linking": "~7.0.3",
+    "expo-router": "^4.0.9",
+    "expo-splash-screen": "~0.29.13",
     "expo-sqlite": "~15.0.3",
     "expo-status-bar": "~2.0.0",
-    "expo-system-ui": "~4.0.2",
+    "expo-system-ui": "~4.0.4",
     "expo-web-browser": "~14.0.1",
     "fix": "^0.0.6",
     "produse": "file:",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-native": "0.76.1",
+    "react-native": "0.76.3",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
-    "react-native-screens": "^4.0.0",
+    "react-native-screens": "~4.1.0",
     "react-native-web": "~0.19.13",
+    "react-navigation": "^5.0.0",
     "sqlite3": "^5.1.7",
     "yarn": "^1.22.22"
   },
@@ -48,7 +53,7 @@
     "@types/react-native": "^0.72.8",
     "@types/react-test-renderer": "^18.0.7",
     "jest": "^29.2.1",
-    "jest-expo": "~52.0.1",
+    "jest-expo": "~52.0.2",
     "react-test-renderer": "18.2.0",
     "typescript": "~5.3.3"
   },
